**2022/09/16** * update function * ProgressDialog

+++ app/src/main/java/com/jack/bookshelf/widget/dialog/PaperProgressDialog.java
@@ -54,7 +54,7 @@ public class PaperProgressDialog extends PopupWindow{
     }
 
     private String getPercentText(int progress) {
-        return (100 * progress / max) + " %";
+        return (int)(100 * progress / max) + " %";
     }
 
     public PaperProgressDialog setButton(int strId) {
